Moves AvroSink to upper level
